+++ src/main/kotlin/app/revanced/patches/music/misc/exclusiveaudio/patch/ExclusiveAudioPatch.kt
@@ -3,18 +3,14 @@ package app.revanced.patches.music.misc.exclusiveaudio.patch
 import app.revanced.extensions.exception
 import app.revanced.patcher.annotation.Description
 import app.revanced.patcher.annotation.Name
-
 import app.revanced.patcher.data.BytecodeContext
-
 import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.patch.BytecodePatch
-import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patcher.patch.annotations.Patch
 import app.revanced.patcher.util.proxy.mutableTypes.MutableMethod
 import app.revanced.patches.music.misc.exclusiveaudio.fingerprints.MusicBrowserServiceFingerprint
 import app.revanced.patches.music.utils.annotations.MusicCompatibility
-import app.revanced.patches.music.utils.fix.decoding.patch.DecodingPatch
 import app.revanced.util.bytecode.getStringIndex
 import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
@@ -22,9 +18,7 @@ import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
 @Patch
 @Name("Exclusive audio playback")
 @Description("Enables the option to play music without video.")
-@DependsOn([DecodingPatch::class])
 @MusicCompatibility
-
 class ExclusiveAudioPatch : BytecodePatch(
     listOf(MusicBrowserServiceFingerprint)
 ) {
@@ -32,7 +26,8 @@ class ExclusiveAudioPatch : BytecodePatch(
 
         MusicBrowserServiceFingerprint.result?.let {
             it.mutableMethod.apply {
-                val targetIndex = getStringIndex("MBS: Return empty root for client: %s, isFullMediaBrowserEnabled: %b, is client browsable: %b, isRedAccount: %b")
+                val targetIndex =
+                    getStringIndex("MBS: Return empty root for client: %s, isFullMediaBrowserEnabled: %b, is client browsable: %b, isRedAccount: %b")
 
                 for (index in targetIndex downTo 0) {
                     if (getInstruction(index).opcode != Opcode.INVOKE_VIRTUAL) continue
@@ -58,5 +53,6 @@ class ExclusiveAudioPatch : BytecodePatch(
                 }
             }
         } ?: throw MusicBrowserServiceFingerprint.exception
+
     }
 }