+++ src/main/kotlin/app/revanced/patches/music/misc/premium/patch/HideGetPremiumPatch.kt
@@ -3,19 +3,20 @@ package app.revanced.patches.music.misc.premium.patch
 import app.revanced.extensions.exception
 import app.revanced.patcher.annotation.Description
 import app.revanced.patcher.annotation.Name
-
 import app.revanced.patcher.data.BytecodeContext
-
 import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
+import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint.Companion.resolve
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patcher.patch.annotations.Patch
 import app.revanced.patcher.util.proxy.mutableTypes.MutableMethod
 import app.revanced.patches.music.misc.premium.fingerprints.AccountMenuFooterFingerprint
 import app.revanced.patches.music.misc.premium.fingerprints.HideGetPremiumFingerprint
+import app.revanced.patches.music.misc.premium.fingerprints.MembershipSettingsFingerprint
+import app.revanced.patches.music.misc.premium.fingerprints.MembershipSettingsParentFingerprint
 import app.revanced.patches.music.utils.annotations.MusicCompatibility
-import app.revanced.patches.music.utils.fix.decoding.patch.DecodingPatch
 import app.revanced.patches.music.utils.resourceid.patch.SharedResourceIdPatch
 import app.revanced.patches.music.utils.resourceid.patch.SharedResourceIdPatch.Companion.PrivacyTosFooter
 import app.revanced.util.bytecode.getWideLiteralIndex
@@ -27,19 +28,14 @@ import com.android.tools.smali.dexlib2.iface.reference.Reference
 
 @Patch
 @Name("Hide get premium")
-@Description("Hides \"Get Premium\" label from the account menu.")
-@DependsOn(
-    [
-        DecodingPatch::class,
-        SharedResourceIdPatch::class
-    ]
-)
+@Description("Hides \"Get Premium\" label from the account menu or settings.")
+@DependsOn([SharedResourceIdPatch::class])
 @MusicCompatibility
-
 class HideGetPremiumPatch : BytecodePatch(
     listOf(
         AccountMenuFooterFingerprint,
-        HideGetPremiumFingerprint
+        HideGetPremiumFingerprint,
+        MembershipSettingsParentFingerprint
     )
 ) {
     override fun execute(context: BytecodeContext) {
@@ -88,6 +84,25 @@ class HideGetPremiumPatch : BytecodePatch(
                 }
             }
         } ?: throw AccountMenuFooterFingerprint.exception
+
+        MembershipSettingsParentFingerprint.result?.let { parentResult ->
+            MembershipSettingsFingerprint.also {
+                it.resolve(
+                    context,
+                    parentResult.classDef
+                )
+            }.result?.let {
+                it.mutableMethod.apply {
+                    addInstructions(
+                        0, """
+                            const/4 v0, 0x0
+                            return-object v0
+                            """
+                    )
+                }
+            } ?: throw MembershipSettingsFingerprint.exception
+        } ?: throw MembershipSettingsParentFingerprint.exception
+
     }
 
     private companion object {