+++ src/main/kotlin/app/revanced/patches/music/utils/resourceid/patch/SharedResourceIdPatch.kt
@@ -1,8 +1,6 @@
 package app.revanced.patches.music.utils.resourceid.patch
 
 import app.revanced.patcher.data.ResourceContext
-import app.revanced.patcher.patch.PatchException
-
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patches.shared.patch.mapping.ResourceMappingPatch
@@ -18,15 +16,19 @@ import app.revanced.util.enum.ResourceType.STYLE
 @DependsOn([ResourceMappingPatch::class])
 class SharedResourceIdPatch : ResourcePatch {
     internal companion object {
+        var ActionsContainer: Long = -1
+        var ButtonIconPaddingMedium: Long = -1
         var ChipCloud: Long = -1
         var ColorGrey: Long = -1
         var DialogSolid: Long = -1
         var DisabledIconAlpha: Long = -1
         var IsTablet: Long = -1
+        var MusicMenuLikeButtons: Long = -1
         var MusicNotifierShelf: Long = -1
         var PrivacyTosFooter: Long = -1
         var QualityTitle: Long = -1
         var Text1: Long = -1
+        var ToolTipContentView: Long = -1
     }
 
     override fun execute(context: ResourceContext) {
@@ -34,16 +36,21 @@ class SharedResourceIdPatch : ResourcePatch {
         fun find(resourceType: ResourceType, resourceName: String) = ResourceMappingPatch
             .resourceMappings
             .find { it.type == resourceType.value && it.name == resourceName }?.id
-            ?: throw PatchException("Failed to find resource id : $resourceName")
+            ?: -1
 
+        ActionsContainer = find(ID, "actions_container")
+        ButtonIconPaddingMedium = find(DIMEN, "button_icon_padding_medium")
         ChipCloud = find(LAYOUT, "chip_cloud")
         ColorGrey = find(COLOR, "ytm_color_grey_12")
         DialogSolid = find(STYLE, "Theme.YouTubeMusic.Dialog.Solid")
         DisabledIconAlpha = find(DIMEN, "disabled_icon_alpha")
         IsTablet = find(BOOL, "is_tablet")
+        MusicMenuLikeButtons = find(LAYOUT, "music_menu_like_buttons")
         MusicNotifierShelf = find(LAYOUT, "music_notifier_shelf")
         PrivacyTosFooter = find(ID, "privacy_tos_footer")
         QualityTitle = find(STRING, "quality_title")
         Text1 = find(ID, "text1")
+        ToolTipContentView = find(LAYOUT, "tooltip_content_view")
+
     }
 }