+++ src/main/kotlin/app/revanced/patches/reddit/utils/resourceid/patch/SharedResourceIdPatch.kt
@@ -1,7 +1,6 @@
 package app.revanced.patches.reddit.utils.resourceid.patch
 
 import app.revanced.patcher.data.ResourceContext
-import app.revanced.patcher.patch.PatchException
 
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotations.DependsOn
@@ -20,7 +19,7 @@ class SharedResourceIdPatch : ResourcePatch {
         fun find(resourceType: ResourceType, resourceName: String) = ResourceMappingPatch
             .resourceMappings
             .find { it.type == resourceType.value && it.name == resourceName }?.id
-            ?: throw PatchException("Failed to find resource id : $resourceName")
+            ?: -1
 
         ScreenShotShareBanner = find(STRING, "screenshot_share_banner_title")
     }