+++ src/main/kotlin/app/revanced/patches/youtube/seekbar/timestamps/fingerprints/TimeCounterFingerprint.kt
@@ -10,26 +10,8 @@ object TimeCounterFingerprint : MethodFingerprint(
     parameters = emptyList(),
     returnType = "V",
     opcodes = listOf(
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.MOVE_RESULT_WIDE,
-        Opcode.INVOKE_STATIC,
-        Opcode.MOVE_RESULT_OBJECT,
-        Opcode.INVOKE_INTERFACE,
-        Opcode.RETURN_VOID,
-        Opcode.IGET_BOOLEAN,
-        Opcode.IF_EQZ,
-        Opcode.IGET_OBJECT,
-        Opcode.IGET_WIDE,
-        Opcode.IGET_WIDE,
         Opcode.SUB_LONG_2ADDR,
-        Opcode.INVOKE_STATIC,
-        Opcode.MOVE_RESULT_OBJECT,
-        null, //iget-wide or iget-object
-        Opcode.IGET_WIDE
-    ),
-    customFingerprint = { _, classDef ->
-        // On older devices this fingerprint resolves very slowly.
-        // Speed this up by checking for the number of methods.
-        classDef.methods.count() == 14 || classDef.methods.count() == 15 || classDef.methods.count() == 17
-    }
+        Opcode.IGET_WIDE,
+        Opcode.SUB_LONG_2ADDR
+    )
 )