+++ src/main/kotlin/app/revanced/patches/youtube/utils/playercontrols/patch/PlayerControlsPatch.kt
@@ -8,7 +8,6 @@ import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint.Companion.
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprintResult
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchException
-
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patcher.util.proxy.mutableTypes.MutableMethod
 import app.revanced.patches.youtube.utils.fingerprints.YouTubeControlsOverlayFingerprint
@@ -18,15 +17,16 @@ import app.revanced.patches.youtube.utils.playercontrols.fingerprints.Fullscreen
 import app.revanced.patches.youtube.utils.playercontrols.fingerprints.FullscreenEngagementSpeedEduVisibleParentFingerprint
 import app.revanced.patches.youtube.utils.playercontrols.fingerprints.PlayerControlsVisibilityFingerprint
 import app.revanced.patches.youtube.utils.playercontrols.fingerprints.PlayerControlsVisibilityModelFingerprint
+import app.revanced.patches.youtube.utils.playercontrols.fingerprints.QuickSeekVisibleFingerprint
 import app.revanced.patches.youtube.utils.playercontrols.fingerprints.SeekEDUVisibleFingerprint
 import app.revanced.patches.youtube.utils.playercontrols.fingerprints.UserScrubbingFingerprint
-import app.revanced.patches.youtube.utils.playercontrols.fingerprints.QuickSeekVisibleFingerprint
 import app.revanced.patches.youtube.utils.resourceid.patch.SharedResourceIdPatch
 import app.revanced.util.bytecode.getStringIndex
 import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.TwoRegisterInstruction
+import com.android.tools.smali.dexlib2.iface.instruction.formats.Instruction35c
 import com.android.tools.smali.dexlib2.iface.reference.Reference
 
 @DependsOn([SharedResourceIdPatch::class])
@@ -43,11 +43,34 @@ class PlayerControlsPatch : BytecodePatch(
 
         fun MutableMethod.findReference(targetString: String): Reference {
             val targetIndex = getStringIndex(targetString) + 2
+            val targetOpcode = getInstruction(targetIndex).opcode
 
-            return getInstruction<ReferenceInstruction>(targetIndex).reference
+            if (targetOpcode == Opcode.INVOKE_VIRTUAL) {
+                val targetRegister = getInstruction<Instruction35c>(targetIndex).registerD
+
+                val instructions = implementation!!.instructions
+                for ((index, instruction) in instructions.withIndex()) {
+                    if (instruction.opcode != Opcode.IGET_BOOLEAN) continue
+
+                    if (getInstruction<TwoRegisterInstruction>(index).registerA == targetRegister)
+                        return getInstruction<ReferenceInstruction>(index).reference
+                }
+            } else if (targetOpcode == Opcode.IGET_BOOLEAN) {
+                return getInstruction<ReferenceInstruction>(targetIndex).reference
+            }
+
+            throw PatchException("Reference not found: $targetString")
         }
 
         PlayerControlsVisibilityModelFingerprint.result?.classDef?.let { classDef ->
+            quickSeekVisibleMutableMethod =
+                QuickSeekVisibleFingerprint.also {
+                    it.resolve(
+                        context,
+                        classDef
+                    )
+                }.result?.mutableMethod ?: throw QuickSeekVisibleFingerprint.exception
+
             seekEDUVisibleMutableMethod =
                 SeekEDUVisibleFingerprint.also {
                     it.resolve(
@@ -63,14 +86,6 @@ class PlayerControlsPatch : BytecodePatch(
                         classDef
                     )
                 }.result?.mutableMethod ?: throw UserScrubbingFingerprint.exception
-            
-            QuickSeekVisibleMutableMethod =
-                QuickSeekVisibleFingerprint.also {
-                    it.resolve(
-                        context,
-                        classDef
-                    )
-                }.result?.mutableMethod ?: throw QuickSeekVisibleFingerprint.exception
         } ?: throw PlayerControlsVisibilityModelFingerprint.exception
 
         YouTubeControlsOverlayFingerprint.result?.classDef?.let { classDef ->
@@ -108,6 +123,7 @@ class PlayerControlsPatch : BytecodePatch(
                 }.result?.mutableMethod
                     ?: throw FullscreenEngagementSpeedEduVisibleFingerprint.exception
         } ?: throw FullscreenEngagementSpeedEduVisibleParentFingerprint.exception
+
     }
 
     internal companion object {
@@ -115,10 +131,10 @@ class PlayerControlsPatch : BytecodePatch(
         lateinit var inflateResult: MethodFingerprintResult
 
         lateinit var playerControlsVisibilityMutableMethod: MutableMethod
+        lateinit var quickSeekVisibleMutableMethod: MutableMethod
         lateinit var seekEDUVisibleMutableMethod: MutableMethod
         lateinit var userScrubbingMutableMethod: MutableMethod
-        lateinit var QuickSeekVisibleMutableMethod: MutableMethod
-        
+
         lateinit var fullscreenEngagementSpeedEduVisibleMutableMethod: MutableMethod
         lateinit var fullscreenEngagementViewVisibleReference: Reference
         lateinit var speedEDUVisibleReference: Reference
@@ -172,6 +188,10 @@ class PlayerControlsPatch : BytecodePatch(
                 descriptor,
                 "changeVisibility"
             )
+            quickSeekVisibleMutableMethod.injectVisibilityCall(
+                descriptor,
+                "changeVisibilityNegatedImmediate"
+            )
             seekEDUVisibleMutableMethod.injectVisibilityCall(
                 descriptor,
                 "changeVisibilityNegatedImmediate"
@@ -180,11 +200,7 @@ class PlayerControlsPatch : BytecodePatch(
                 descriptor,
                 "changeVisibilityNegatedImmediate"
             )
-            QuickSeekVisibleMutableMethod.injectVisibilityCall(
-                descriptor,
-                "changeVisibilityNegatedImmediate"
-            )
-            
+
             injectFullscreenEngagementSpeedEduViewVisibilityCall(
                 fullscreenEngagementViewVisibleReference,
                 descriptor