+++ src/main/kotlin/app/revanced/patches/youtube/utils/resourceid/patch/SharedResourceIdPatch.kt
@@ -1,8 +1,6 @@
 package app.revanced.patches.youtube.utils.resourceid.patch
 
 import app.revanced.patcher.data.ResourceContext
-import app.revanced.patcher.patch.PatchException
-
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patches.shared.patch.mapping.ResourceMappingPatch
@@ -30,7 +28,6 @@ class SharedResourceIdPatch : ResourcePatch {
         var Bar: Long = -1
         var BarContainerHeight: Long = -1
         var BottomPanelOverlayText: Long = -1
-        var BottomSheetMargins: Long = -1
         var BottomUiContainerStub: Long = -1
         var ChannelListSubMenu: Long = -1
         var CompactLink: Long = -1
@@ -60,6 +57,7 @@ class SharedResourceIdPatch : ResourcePatch {
         var ReelDynShare: Long = -1
         var ReelPlayerBadge: Long = -1
         var ReelPlayerBadge2: Long = -1
+        var ReelPlayerFooter: Long = -1
         var ReelPlayerInfoPanel: Long = -1
         var ReelPlayerPausedStateButton: Long = -1
         var ReelRightDislikeIcon: Long = -1
@@ -79,6 +77,7 @@ class SharedResourceIdPatch : ResourcePatch {
         var VideoZoomIndicatorLayout: Long = -1
         var WordMarkHeader: Long = -1
         var YoutubeControlsOverlay: Long = -1
+        var YtBrandBackgroundSolid: Long = -1
         var YtOutlineArrowTimeBlack: Long = -1
         var YtOutlineFireBlack: Long = -1
         var YtOutlineSearchBlack: Long = -1
@@ -89,7 +88,7 @@ class SharedResourceIdPatch : ResourcePatch {
         fun find(resourceType: ResourceType, resourceName: String) = ResourceMappingPatch
             .resourceMappings
             .find { it.type == resourceType.value && it.name == resourceName }?.id
-            ?: throw PatchException("Failed to find resource id : $resourceName")
+            ?: -1
 
         AccountSwitcherAccessibility = find(STRING, "account_switcher_accessibility_label")
         AccessibilityCaptionsButtonName = find(STRING, "accessibility_captions_button_name")
@@ -103,7 +102,6 @@ class SharedResourceIdPatch : ResourcePatch {
         Bar = find(LAYOUT, "bar")
         BarContainerHeight = find(DIMEN, "bar_container_height")
         BottomPanelOverlayText = find(ID, "bottom_panel_overlay_text")
-        BottomSheetMargins = find(DIMEN, "bottom_sheet_margins")
         BottomUiContainerStub = find(ID, "bottom_ui_container_stub")
         ChannelListSubMenu = find(LAYOUT, "channel_list_sub_menu")
         CompactLink = find(LAYOUT, "compact_link")
@@ -135,6 +133,7 @@ class SharedResourceIdPatch : ResourcePatch {
         ReelDynShare = find(ID, "reel_dyn_share")
         ReelPlayerBadge = find(ID, "reel_player_badge")
         ReelPlayerBadge2 = find(ID, "reel_player_badge2")
+        ReelPlayerFooter = find(LAYOUT, "reel_player_dyn_footer_vert_stories3")
         ReelPlayerInfoPanel = find(ID, "reel_player_info_panel")
         ReelPlayerPausedStateButton = find(ID, "reel_player_paused_state_buttons")
         ReelRightDislikeIcon = find(DRAWABLE, "reel_right_dislike_icon")
@@ -154,8 +153,10 @@ class SharedResourceIdPatch : ResourcePatch {
         VideoZoomIndicatorLayout = find(ID, "video_zoom_indicator_layout")
         WordMarkHeader = find(ATTR, "ytWordmarkHeader")
         YoutubeControlsOverlay = find(ID, "youtube_controls_overlay")
+        YtBrandBackgroundSolid = find(ATTR, "ytBrandBackgroundSolid")
         YtOutlineArrowTimeBlack = find(DRAWABLE, "yt_outline_arrow_time_black_24")
         YtOutlineFireBlack = find(DRAWABLE, "yt_outline_fire_black_24")
         YtOutlineSearchBlack = find(DRAWABLE, "yt_outline_search_black_24")
+
     }
 }