+++ src/main/kotlin/app/revanced/patches/youtube/utils/returnyoutubedislike/shorts/fingerprints/ShortsTextViewFingerprint.kt
@@ -24,7 +24,5 @@ object ShortsTextViewFingerprint : MethodFingerprint(
         Opcode.IF_EQ,
         Opcode.RETURN_VOID,
         Opcode.IGET_OBJECT,     // TextView field
-        Opcode.IGET_BOOLEAN,    // boolean field
-        Opcode.CHECK_CAST
     )
 )